requirement.txt
fastapi==0.135.1
numpy==2.3.5
pydantic==2.12.5
python-multipart==0.0.22
PyYAML==6.0.3
torch==2.10.0+cu126
torchvision==0.25.0+cu126
uvicorn==0.41.0
